GUI Application/requirements.txt
pyqtgraph==0.12.3
numpy==1.19.5
PyQt5==5.15.6
pyserial==3.5
